@@ -83,7 +83,13 @@ COMMANDS = [
     "plugin remove",
     "plugin show",
     # Self commands
+    "self add",
+    "self install",
+    "self lock",
+    "self remove",
     "self update",
+    "self show",
+    "self show plugins",
     # Source commands
     "source add",
     "source remove",
@@ -104,7 +110,7 @@ class Application(BaseApplication):  # type: ignore[misc]
         dispatcher = EventDispatcher()
         dispatcher.add_listener(COMMAND, self.register_command_loggers)
         dispatcher.add_listener(COMMAND, self.configure_env)
-        dispatcher.add_listener(COMMAND, self.configure_installer)
+        dispatcher.add_listener(COMMAND, self.configure_installer_for_event)
         self.set_event_dispatcher(dispatcher)
 
         command_loader = CommandLoader({name: load_command(name) for name in COMMANDS})
@@ -296,8 +302,9 @@ class Application(BaseApplication):  # type: ignore[misc]
 
         command.set_env(env)
 
-    def configure_installer(
-        self, event: ConsoleCommandEvent, event_name: str, _: Any
+    @classmethod
+    def configure_installer_for_event(
+        cls, event: ConsoleCommandEvent, event_name: str, _: Any
     ) -> None:
         from poetry.console.commands.installer_command import InstallerCommand
 
@@ -310,9 +317,10 @@ class Application(BaseApplication):  # type: ignore[misc]
         if command.installer is not None:
             return
 
-        self._configure_installer(command, event.io)
+        cls.configure_installer_for_command(command, event.io)
 
-    def _configure_installer(self, command: InstallerCommand, io: IO) -> None:
+    @staticmethod
+    def configure_installer_for_command(command: InstallerCommand, io: IO) -> None:
         from poetry.installation.installer import Installer
 
         poetry = command.poetry
