@@ -25,5 +25,5 @@ class DebugInfoCommand(Command):
         )
         command = self.application.get("env info")
 
-        exit_code: int = command.run(self._io)
+        exit_code: int = command.run(self.io)
         return exit_code
