@@ -12,7 +12,7 @@ class EnvListCommand(Command):
 
     options = [option("full-path", None, "Output the full paths of the virtualenvs.")]
 
-    def handle(self) -> None:
+    def handle(self) -> int:
         from poetry.utils.env import EnvManager
 
         manager = EnvManager(self.poetry)
@@ -29,3 +29,5 @@ class EnvListCommand(Command):
                 continue
 
             self.line(name)
+
+        return 0
