@@ -29,7 +29,7 @@ class EnvRemoveCommand(Command):
         ),
     ]
 
-    def handle(self) -> None:
+    def handle(self) -> int:
         from poetry.utils.env import EnvManager
 
         pythons = self.argument("python")
@@ -46,3 +46,5 @@ class EnvRemoveCommand(Command):
             for venv in manager.list():
                 manager.remove_venv(venv.path)
                 self.line(f"Deleted virtualenv: <comment>{venv.path}</comment>")
+
+        return 0
