@@ -4,17 +4,11 @@ from typing import Any
 
 from cleo.helpers import argument
 from cleo.helpers import option
+from poetry.core.packages.dependency_group import MAIN_GROUP
 from tomlkit.toml_document import TOMLDocument
 
-from poetry.utils.helpers import canonicalize_name
-
-
-try:
-    from poetry.core.packages.dependency_group import MAIN_GROUP
-except ImportError:
-    MAIN_GROUP = "default"
-
 from poetry.console.commands.installer_command import InstallerCommand
+from poetry.utils.helpers import canonicalize_name
 
 
 class RemoveCommand(InstallerCommand):
@@ -81,12 +75,17 @@ list of installed packages
             if not poetry_content["dev-dependencies"]:
                 del poetry_content["dev-dependencies"]
         else:
-            removed = self._remove_packages(
-                packages, poetry_content["group"][group].get("dependencies", {}), group
-            )
+            removed = []
+            if "group" in poetry_content:
+                if group in poetry_content["group"]:
+                    removed = self._remove_packages(
+                        packages,
+                        poetry_content["group"][group].get("dependencies", {}),
+                        group,
+                    )
 
-            if not poetry_content["group"][group]:
-                del poetry_content["group"][group]
+                if not poetry_content["group"][group]:
+                    del poetry_content["group"][group]
 
         if "group" in poetry_content and not poetry_content["group"]:
             del poetry_content["group"]
@@ -104,13 +103,9 @@ list of installed packages
         )
         self._installer.set_locker(self.poetry.locker)
 
-        # Update packages
-        self._installer.use_executor(
-            self.poetry.config.get("experimental.new-installer", False)
-        )
-
+        self._installer.set_package(self.poetry.package)
         self._installer.dry_run(self.option("dry-run", False))
-        self._installer.verbose(self._io.is_verbose())
+        self._installer.verbose(self.io.is_verbose())
         self._installer.update(True)
         self._installer.whitelist(removed_set)
 
