@@ -12,7 +12,7 @@ class SearchCommand(Command):
 
     arguments = [argument("tokens", "The tokens to search for.", multiple=True)]
 
-    def handle(self) -> None:
+    def handle(self) -> int:
         from poetry.repositories.pypi_repository import PyPiRepository
 
         results = PyPiRepository().search(self.argument("tokens"))
@@ -27,3 +27,5 @@ class SearchCommand(Command):
 
             if result.description:
                 self.line(f" {result.description}")
+
+        return 0
