@@ -1,7 +1,6 @@
 from __future__ import annotations
 
 from typing import TYPE_CHECKING
-from typing import Any
 
 from poetry.mixology.term import Term
 
@@ -51,7 +50,7 @@ class Assignment(Term):
     @classmethod
     def derivation(
         cls,
-        dependency: Any,
+        dependency: Dependency,
         is_positive: bool,
         cause: Incompatibility,
         decision_level: int,
