@@ -1,6 +1,7 @@
 from __future__ import annotations
 
 from typing import TYPE_CHECKING
+from typing import List
 
 from poetry.packages.dependency_package import DependencyPackage
 
@@ -12,7 +13,7 @@ if TYPE_CHECKING:
     from poetry.core.packages.package import Package
 
 
-class PackageCollection(list):  # type: ignore[type-arg]
+class PackageCollection(List[DependencyPackage]):
     def __init__(
         self,
         dependency: Dependency,
