@@ -33,7 +33,7 @@ class HTMLPage(LinkSource):
                 url = self.clean_link(urllib.parse.urljoin(self._url, href))
                 pyrequire = anchor.get("data-requires-python")
                 pyrequire = unescape(pyrequire) if pyrequire else None
-                link = Link(url, self, requires_python=pyrequire)
+                link = Link(url, requires_python=pyrequire)
 
                 if link.ext not in self.SUPPORTED_FORMATS:
                     continue
