@@ -5,8 +5,10 @@ import sys
 from contextlib import suppress
 
 
-if sys.version_info < (3, 8):
-    # compatibility for python <3.8
+# TODO: use try/except ImportError when
+# https://github.com/python/mypy/issues/1393 is fixed
+if sys.version_info < (3, 10):
+    # compatibility for python <3.10
     import importlib_metadata as metadata
 else:
     from importlib import metadata
