@@ -6,6 +6,7 @@ import itertools
 import json
 import os
 import platform
+import plistlib
 import re
 import subprocess
 import sys
@@ -1087,7 +1088,20 @@ class EnvManager:
 
         args.append(str(path))
 
-        return virtualenv.cli_run(args)
+        cli_result = virtualenv.cli_run(args)
+
+        # Exclude the venv folder from from macOS Time Machine backups
+        # TODO: Add backup-ignore markers for other platforms too
+        if sys.platform == "darwin":
+            import xattr
+
+            xattr.setxattr(
+                str(path),
+                "com.apple.metadata:com_apple_backup_excludeItem",
+                plistlib.dumps("com.apple.backupd", fmt=plistlib.FMT_BINARY),
+            )
+
+        return cli_result
 
     @classmethod
     def remove_venv(cls, path: Path | str) -> None:
@@ -1484,13 +1498,14 @@ class Env:
 
         return decode(output)
 
-    def execute(self, bin: str, *args: str, **kwargs: Any) -> int | None:
+    def execute(self, bin: str, *args: str, **kwargs: Any) -> int:
         command = self.get_command_from_bin(bin) + list(args)
         env = kwargs.pop("env", dict(os.environ))
 
         if not self._is_windows:
             return os.execvpe(command[0], command, env=env)
 
+        kwargs["shell"] = True
         exe = subprocess.Popen([command[0]] + command[1:], env=env, **kwargs)
         exe.communicate()
         return exe.returncode
@@ -1753,7 +1768,7 @@ class VirtualEnv(Env):
 
         return environ
 
-    def execute(self, bin: str, *args: str, **kwargs: Any) -> int | None:
+    def execute(self, bin: str, *args: str, **kwargs: Any) -> int:
         kwargs["env"] = self.get_temp_environ(environ=kwargs.get("env"))
         return super().execute(bin, *args, **kwargs)
 
@@ -1836,7 +1851,7 @@ class GenericEnv(VirtualEnv):
         paths: dict[str, str] = json.loads(output)
         return paths
 
-    def execute(self, bin: str, *args: str, **kwargs: Any) -> int | None:
+    def execute(self, bin: str, *args: str, **kwargs: Any) -> int:
         command = self.get_command_from_bin(bin) + list(args)
         env = kwargs.pop("env", dict(os.environ))
 
@@ -1880,12 +1895,12 @@ class NullEnv(SystemEnv):
             return super()._run(cmd, **kwargs)
         return 0
 
-    def execute(self, bin: str, *args: str, **kwargs: Any) -> int | None:
+    def execute(self, bin: str, *args: str, **kwargs: Any) -> int:
         self.executed.append([bin] + list(args))
 
         if self._execute:
             return super().execute(bin, *args, **kwargs)
-        return None
+        return 0
 
     def _bin(self, bin: str) -> str:
         return bin
