@@ -36,9 +36,7 @@ def test_publish_returns_non_zero_code_for_upload_errors(
 Publishing simple-project (1.2.3) to PyPI
 """
     expected_error_output = """\
-  UploadError
-
-  HTTP Error 400: Bad Request
+HTTP Error 400: Bad Request
 """
 
     assert expected_output in app_tester.io.fetch_output()
